# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for retrieving time series data from stock markets",
        "title": "Stock Market API",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/fundamental/balance-sheet/{symbol}": {
            "get": {
                "description": "Returns the balance sheet data for the specified stock symbol",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fundamental"
                ],
                "summary": "Get balance sheet data for a specific symbol",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stock symbol (e.g., AAPL, MSFT)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/alphavantage.BalanceSheetResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/fundamental/cash-flow/{symbol}": {
            "get": {
                "description": "Returns the cash flow data for the specified stock symbol",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fundamental"
                ],
                "summary": "Get cash flow data for a specific symbol",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stock symbol (e.g., AAPL, MSFT)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/alphavantage.CashFlowResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/fundamental/company-overview/{symbol}": {
            "get": {
                "description": "Returns the company overview data for the specified stock symbol (sector, industry, PE ratio, EBITDA, and more)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fundamental"
                ],
                "summary": "Get company overview data for a specific symbol",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stock symbol (e.g., AAPL, MSFT)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/alphavantage.CompanyOverviewResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/fundamental/income-statement/{symbol}": {
            "get": {
                "description": "Returns the income statement data for the specified stock symbol",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fundamental"
                ],
                "summary": "Get income statement data for a specific symbol",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stock symbol (e.g., AAPL, MSFT)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/alphavantage.IncomeStatementResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "alphavantage.BalanceSheetResponse": {
            "description": "Balance sheet response data structure",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/fundamental.BalanceSheetResponse"
                },
                "symbol": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "alphavantage.CashFlowResponse": {
            "description": "Cash flow response data structure",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/fundamental.CashFlowResponse"
                },
                "symbol": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "alphavantage.CompanyOverviewResponse": {
            "description": "Company overview response data structure",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/fundamental.CompanyOverviewResponse"
                },
                "symbol": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "alphavantage.IncomeStatementResponse": {
            "description": "Income statement response data structure",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/fundamental.IncomeStatementResponse"
                },
                "symbol": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "fundamental.BalanceSheetReport": {
            "type": "object",
            "properties": {
                "accumulatedDepreciationAmortizationPPE": {
                    "type": "string"
                },
                "capitalLeaseObligations": {
                    "type": "string"
                },
                "cashAndCashEquivalentsAtCarryingValue": {
                    "type": "string"
                },
                "cashAndShortTermInvestments": {
                    "type": "string"
                },
                "commonStock": {
                    "type": "string"
                },
                "commonStockSharesOutstanding": {
                    "type": "string"
                },
                "currentAccountsPayable": {
                    "type": "string"
                },
                "currentDebt": {
                    "type": "string"
                },
                "currentLongTermDebt": {
                    "type": "string"
                },
                "currentNetReceivables": {
                    "type": "string"
                },
                "deferredRevenue": {
                    "type": "string"
                },
                "fiscalDateEnding": {
                    "type": "string"
                },
                "goodwill": {
                    "type": "string"
                },
                "intangibleAssets": {
                    "type": "string"
                },
                "intangibleAssetsExcludingGoodwill": {
                    "type": "string"
                },
                "inventory": {
                    "type": "string"
                },
                "investments": {
                    "type": "string"
                },
                "longTermDebt": {
                    "type": "string"
                },
                "longTermDebtNoncurrent": {
                    "type": "string"
                },
                "longTermInvestments": {
                    "type": "string"
                },
                "otherCurrentAssets": {
                    "type": "string"
                },
                "otherCurrentLiabilities": {
                    "type": "string"
                },
                "otherNonCurrentAssets": {
                    "type": "string"
                },
                "otherNonCurrentLiabilities": {
                    "type": "string"
                },
                "propertyPlantEquipment": {
                    "type": "string"
                },
                "reportedCurrency": {
                    "type": "string"
                },
                "retainedEarnings": {
                    "type": "string"
                },
                "shortLongTermDebtTotal": {
                    "type": "string"
                },
                "shortTermDebt": {
                    "type": "string"
                },
                "shortTermInvestments": {
                    "type": "string"
                },
                "totalAssets": {
                    "type": "string"
                },
                "totalCurrentAssets": {
                    "type": "string"
                },
                "totalCurrentLiabilities": {
                    "type": "string"
                },
                "totalLiabilities": {
                    "type": "string"
                },
                "totalNonCurrentAssets": {
                    "type": "string"
                },
                "totalNonCurrentLiabilities": {
                    "type": "string"
                },
                "totalShareholderEquity": {
                    "type": "string"
                },
                "treasuryStock": {
                    "type": "string"
                }
            }
        },
        "fundamental.BalanceSheetResponse": {
            "type": "object",
            "properties": {
                "annualReports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fundamental.BalanceSheetReport"
                    }
                },
                "quarterlyReports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fundamental.BalanceSheetReport"
                    }
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "fundamental.CashFlowReport": {
            "type": "object",
            "properties": {
                "capitalExpenditures": {
                    "type": "string"
                },
                "cashflowFromFinancing": {
                    "type": "string"
                },
                "cashflowFromInvestment": {
                    "type": "string"
                },
                "changeInCashAndCashEquivalents": {
                    "type": "string"
                },
                "changeInExchangeRate": {
                    "type": "string"
                },
                "changeInInventory": {
                    "type": "string"
                },
                "changeInOperatingAssets": {
                    "type": "string"
                },
                "changeInOperatingLiabilities": {
                    "type": "string"
                },
                "changeInReceivables": {
                    "type": "string"
                },
                "depreciationDepletionAndAmortization": {
                    "type": "string"
                },
                "dividendPayout": {
                    "type": "string"
                },
                "dividendPayoutCommonStock": {
                    "type": "string"
                },
                "dividendPayoutPreferredStock": {
                    "type": "string"
                },
                "fiscalDateEnding": {
                    "type": "string"
                },
                "netIncome": {
                    "type": "string"
                },
                "operatingCashflow": {
                    "type": "string"
                },
                "paymentsForOperatingActivities": {
                    "type": "string"
                },
                "paymentsForRepurchaseOfCommonStock": {
                    "type": "string"
                },
                "paymentsForRepurchaseOfEquity": {
                    "type": "string"
                },
                "paymentsForRepurchaseOfPreferredStock": {
                    "type": "string"
                },
                "proceedsFromIssuanceOfCommonStock": {
                    "type": "string"
                },
                "proceedsFromIssuanceOfLongTermDebtAndCapitalSecuritiesNet": {
                    "type": "string"
                },
                "proceedsFromIssuanceOfPreferredStock": {
                    "type": "string"
                },
                "proceedsFromOperatingActivities": {
                    "type": "string"
                },
                "proceedsFromRepaymentsOfShortTermDebt": {
                    "type": "string"
                },
                "proceedsFromRepurchaseOfEquity": {
                    "type": "string"
                },
                "proceedsFromSaleOfTreasuryStock": {
                    "type": "string"
                },
                "profitLoss": {
                    "type": "string"
                },
                "reportedCurrency": {
                    "type": "string"
                }
            }
        },
        "fundamental.CashFlowResponse": {
            "type": "object",
            "properties": {
                "annualReports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fundamental.CashFlowReport"
                    }
                },
                "quarterlyReports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fundamental.CashFlowReport"
                    }
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "fundamental.CompanyOverviewResponse": {
            "type": "object",
            "properties": {
                "200DayMovingAverage": {
                    "type": "string"
                },
                "50DayMovingAverage": {
                    "type": "string"
                },
                "52WeekHigh": {
                    "type": "string"
                },
                "52WeekLow": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "analystTargetPrice": {
                    "type": "string"
                },
                "beta": {
                    "type": "string"
                },
                "bookValue": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dilutedEPSTTM": {
                    "type": "string"
                },
                "dividendDate": {
                    "type": "string"
                },
                "dividendPerShare": {
                    "type": "string"
                },
                "dividendYield": {
                    "type": "string"
                },
                "ebitda": {
                    "type": "string"
                },
                "eps": {
                    "type": "string"
                },
                "evToEBITDA": {
                    "type": "string"
                },
                "evToRevenue": {
                    "type": "string"
                },
                "exDividendDate": {
                    "type": "string"
                },
                "exchange": {
                    "type": "string"
                },
                "fiscalYearEnd": {
                    "type": "string"
                },
                "forwardAnnualDividendRate": {
                    "type": "string"
                },
                "forwardAnnualDividendYield": {
                    "type": "string"
                },
                "forwardPE": {
                    "type": "string"
                },
                "grossProfitTTM": {
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "lastSplitDate": {
                    "type": "string"
                },
                "lastSplitFactor": {
                    "type": "string"
                },
                "latestQuarter": {
                    "type": "string"
                },
                "marketCapitalization": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "operatingMarginTTM": {
                    "type": "string"
                },
                "payoutRatio": {
                    "type": "string"
                },
                "peRatio": {
                    "type": "string"
                },
                "pegRatio": {
                    "type": "string"
                },
                "percentInsiders": {
                    "type": "string"
                },
                "percentInstitutions": {
                    "type": "string"
                },
                "priceToBookRatio": {
                    "type": "string"
                },
                "priceToSalesRatioTTM": {
                    "type": "string"
                },
                "profitMargin": {
                    "type": "string"
                },
                "quarterlyEarningsGrowthYOY": {
                    "type": "string"
                },
                "quarterlyRevenueGrowthYOY": {
                    "type": "string"
                },
                "returnOnAssetsTTM": {
                    "type": "string"
                },
                "returnOnEquityTTM": {
                    "type": "string"
                },
                "revenuePerShareTTM": {
                    "type": "string"
                },
                "revenueTTM": {
                    "type": "string"
                },
                "sector": {
                    "type": "string"
                },
                "sharesFloat": {
                    "type": "string"
                },
                "sharesOutstanding": {
                    "type": "string"
                },
                "sharesShort": {
                    "type": "string"
                },
                "sharesShortPriorMonth": {
                    "type": "string"
                },
                "shortPercentFloat": {
                    "type": "string"
                },
                "shortPercentOutstanding": {
                    "type": "string"
                },
                "shortRatio": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "trailingPE": {
                    "type": "string"
                }
            }
        },
        "fundamental.IncomeStatementReport": {
            "type": "object",
            "properties": {
                "comprehensiveIncomeNetOfTax": {
                    "type": "string"
                },
                "costOfRevenue": {
                    "type": "string"
                },
                "costofGoodsAndServicesSold": {
                    "type": "string"
                },
                "depreciation": {
                    "type": "string"
                },
                "depreciationAndAmortization": {
                    "type": "string"
                },
                "ebit": {
                    "type": "string"
                },
                "ebitda": {
                    "type": "string"
                },
                "fiscalDateEnding": {
                    "type": "string"
                },
                "grossProfit": {
                    "type": "string"
                },
                "incomeBeforeTax": {
                    "type": "string"
                },
                "incomeTaxExpense": {
                    "type": "string"
                },
                "interestAndDebtExpense": {
                    "type": "string"
                },
                "interestExpense": {
                    "type": "string"
                },
                "interestIncome": {
                    "type": "string"
                },
                "investmentIncomeNet": {
                    "type": "string"
                },
                "netIncome": {
                    "type": "string"
                },
                "netIncomeFromContinuingOperations": {
                    "type": "string"
                },
                "netInterestIncome": {
                    "type": "string"
                },
                "nonInterestIncome": {
                    "type": "string"
                },
                "operatingExpenses": {
                    "type": "string"
                },
                "operatingIncome": {
                    "type": "string"
                },
                "otherNonOperatingIncome": {
                    "type": "string"
                },
                "reportedCurrency": {
                    "type": "string"
                },
                "researchAndDevelopment": {
                    "type": "string"
                },
                "sellingGeneralAndAdministrative": {
                    "type": "string"
                },
                "totalRevenue": {
                    "type": "string"
                }
            }
        },
        "fundamental.IncomeStatementResponse": {
            "type": "object",
            "properties": {
                "annualReports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fundamental.IncomeStatementReport"
                    }
                },
                "quarterlyReports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fundamental.IncomeStatementReport"
                    }
                },
                "symbol": {
                    "type": "string"
                }
            }
        }
    }
}